Add files via upload
•	In this project, I investigated data from the iOS App “Cookie Cats” 
•	An earlier version of the apps placed the final level at gate 30. 
•	But an updated version added 10 more gates, making the final gate 40.
•	This data contains players who tried the 30-gate and the 40-gate version. So, these gates are the 2 different A/B groups.
•	In this program, I did the following:
o	Cleaned the data
o	Split it into A/B groups based on gate
o	Plotted and analyzed different variables based on gate A/B group.
o	Used the machine learning technique of bootstrapping in order to increase the accuracy of my analysis:
♣	Because I created 500 random samples, a large sample size, my findings are highly likely to be representative to populations of similar size. 
♣	This is, however, not necessarily a thorough test for statistical significance.
o	Performed another A/B test based on is_addicted, which I calculated based on whether the player’s sum of game rounds exceeded the median
•	Findings:
o	The simple bar graphs show that at a surface level, retention is roughly the same for both A/B groups for both 1 and 7 days after download.
o	The proportion of retention after 1 day (around 0.4) is much higher than after 7 days (around 0.2) for both A/B groups.
o	Bootstrap analysis indicates that the difference between both groups in 1-day retention, 7-day retention, and is_addicted is significant, favoring gate_30 results. 
o	This is because the probability that the difference between the gate groups is positive exceeds 95%. This is of course at the alpha=0.05 significance level.
o	The same probability for is_addicted also exceeds 95%.
o	Therefore, it makes more sense that Cookie Cats keep the last gate at 30 since the marginal utility of continuing to play decreases sharply after that gate.